requirements.txt
requests == 2.25.1
PyYaml == 5.3.1